added integration tests, removed oldarch
