+++ src/main/java/dev/fastgql/FastGQL.java
@@ -6,7 +6,9 @@
 
 package dev.fastgql;
 
+import com.google.common.collect.Iterables;
 import dev.fastgql.db.DatabaseSchema;
+import dev.fastgql.db.DatasourceConfig;
 import dev.fastgql.db.MetadataUtils;
 import dev.fastgql.graphql.GraphQLDefinition;
 import dev.fastgql.router.RouterUpdatable;
@@ -17,21 +19,36 @@ import io.vertx.core.Promise;
 import io.vertx.core.http.HttpServerOptions;
 import io.vertx.pgclient.PgConnectOptions;
 import io.vertx.reactivex.core.AbstractVerticle;
+import io.vertx.reactivex.kafka.client.consumer.KafkaConsumer;
 import io.vertx.reactivex.pgclient.PgPool;
 import io.vertx.reactivex.sqlclient.Pool;
 import io.vertx.sqlclient.PoolOptions;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
-import java.util.concurrent.TimeUnit;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
 
 public class FastGQL extends AbstractVerticle {
+
+  private static final String alteredTablesTopic = "altered-tables";
+
   public static void main(String[] args) {
-    Launcher.executeCommand("run", FastGQL.class.getName());
+    Launcher.executeCommand("run", FastGQL.class.getName(), "--conf", "src/main/conf.json");
+  }
+
+  private static String kafkaMessageToTableName(String message) {
+    return Iterables.getLast(
+        Arrays.asList(message.substring(1, message.length() - 1).split("\\.")));
   }
 
   @Override
   public void start(Promise<Void> future) throws SQLException {
+
+    DatasourceConfig datasourceConfig = new DatasourceConfig(config().getJsonObject("datasource"));
 
     Connection connection =
         DriverManager.getConnection(
@@ -43,19 +60,37 @@ public class FastGQL extends AbstractVerticle {
         PgPool.pool(
             vertx,
             new PgConnectOptions()
-                .setHost("localhost")
-                .setPort(5432)
-                .setDatabase("quarkus_test")
-                .setUser("quarkus_test")
-                .setPassword("quarkus_test"),
+                .setHost(datasourceConfig.getHost())
+                .setPort(datasourceConfig.getPort())
+                .setDatabase(datasourceConfig.getDb())
+                .setUser(datasourceConfig.getUsername())
+                .setPassword(datasourceConfig.getPassword()),
             new PoolOptions().setMaxSize(5));
 
-    Flowable<String> fake = Flowable.interval(1, TimeUnit.SECONDS).map(i -> "customers");
+    Map<String, String> kafkaConfig = new HashMap<>();
+    kafkaConfig.put(
+        ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG,
+        config().getString("bootstrap.servers", "http://localhost:9092"));
+    kafkaConfig.put(
+        ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG,
+        "org.apache.kafka.common.serialization.StringDeserializer");
+    kafkaConfig.put(
+        ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG,
+        "org.apache.kafka.common.serialization.StringDeserializer");
+    kafkaConfig.put(ConsumerConfig.GROUP_ID_CONFIG, "tc-" + UUID.randomUUID());
+    kafkaConfig.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+    KafkaConsumer<String, String> kafkaConsumer = KafkaConsumer.create(vertx, kafkaConfig);
+    kafkaConsumer.subscribe(alteredTablesTopic);
+
+    Flowable<String> alteredTablesFlowable =
+        kafkaConsumer.toFlowable().map(record -> kafkaMessageToTableName(record.value()));
+
     GraphQL graphQL =
         GraphQLDefinition.newGraphQL(database, client)
             .enableQuery()
-            .enableSubscription(fake)
+            .enableSubscription(alteredTablesFlowable)
             .build();
+
     RouterUpdatable routerUpdatable = RouterUpdatable.createWithQueryAndSubscription(vertx);
     routerUpdatable.update(graphQL);
