--- src/main/java/dev/fastgql/oldarch/Column.java
@@ -1,17 +1,0 @@
-/*
- * Copyright fastGQL Authors.
- *
- * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package dev.fastgql.oldarch;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class Column<T> {
-
-  private final Class<T> clazz;
-  private final String referenceTable;
-}
