+++ src/main/resources/log4j.properties
@@ -1,4 +1,4 @@
-log4j.rootLogger=info, stdout, R
+log4j.rootLogger=warn, stdout, R
 log4j.logger.dev.fastgql.sql=debug
 
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
